+++ src/app/models/post.model.ts
@@ -1,0 +1,9 @@
+export interface Post {
+    id: number;
+    title: string;
+    alt: string;
+    image: string;
+    excerpt: string;
+    content: string;
+  }
+