+++ src/app/pages/blog-detail/blog-detail.component.ts
@@ -1,0 +1,23 @@
+import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
+import { BlogService } from 'src/app/services/blog.service';
+import { Post } from 'src/app/models/post.model';
+
+@Component({
+  selector: 'app-blog-detail',
+  templateUrl: './blog-detail.component.html',
+  styleUrls: ['./blog-detail.component.scss']
+})
+export class BlogDetailComponent implements OnInit {
+  post: Post | undefined;
+
+  constructor(
+    private route: ActivatedRoute,
+    private blogService: BlogService
+  ) {}
+
+  ngOnInit(): void {
+    const id = Number(this.route.snapshot.paramMap.get('id'));
+    this.post = this.blogService.getPostById(id);
+  }
+}