+++ src/environments/environment.prod.ts
@@ -1,5 +1,6 @@
 export const environment = {
     production: true,
-    version: '1.1.0'
+    version: '0.0.4',
+    emailJsPublicKey: '1n4HalUHTYL22tCyx'
   };